+++ src/api/communication.ts
@@ -13,7 +13,7 @@
 import { APPLICATION_NAME } from './env.ts';
 import { ERRORS_IGNORED } from './const.ts';
 import type { TMetrics } from '../api-monitor-cs-main.ts';
-import type { TTimerType } from './wrappers.ts';
+import type { TTimerType } from '../wrapper/main.ts';
 import type { TSettings } from './settings.ts';
 
 let port: chrome.runtime.Port | null = null;
