+++ src/view/components/TimersClearHistoryMetric.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type { TClearTimerHistory } from '../../api/wrappers.ts';
+  import type { TClearTimerHistory } from '../../wrapper/main.ts';
   import { delayTitle } from '../../api/time.ts';
   import Trace from './Trace.svelte';
   import TraceDomain from './TraceDomain.svelte';
