+++ src/view/components/TimersSetHistoryMetric.svelte
@@ -2,7 +2,7 @@
   import type {
     TClearTimerHistory,
     TSetTimerHistory,
-  } from '../../api/wrappers.ts';
+  } from '../../wrapper/main.ts';
   import { CALLED_ABORTED_TOOLTIP } from '../../api/const.ts';
   import { delayTitle } from '../../api/time.ts';
   import FrameSensitiveTime from './FrameSensitiveTime.svelte';
